Moving entity physical body + position over to use Box2D, for easy collision detection

+++ core/src/com/jdkd/test/ecs/component/PreviousPosition.java
@@ -4,10 +4,10 @@ import com.artemis.Component;
 
 public class PreviousPosition extends Component {
 
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public PreviousPosition(int x, int y) {
+    public PreviousPosition(float x, float y) {
         this.x = x;
         this.y = y;
     }
@@ -15,19 +15,19 @@ public class PreviousPosition extends Component {
     public PreviousPosition() {
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
